requirements.txt
beautifulsoup4
flask
flask_cors
python-dotenv
openpyxl
pandas
PyPDF2
python-docx
duckdb
openai
google-auth
google-auth-oauthlib
google-api-python-client
gunicorn